+++ packages/brick_graphql_generators/lib/src/graphql_deserialize.dart
@@ -1,5 +1,5 @@
 import 'package:analyzer/dart/element/element.dart';
-import 'package:brick_graphql/graphql.dart';
+import 'package:brick_graphql/brick_graphql.dart';
 import 'package:brick_graphql_generators/src/graphql_fields.dart';
 import 'package:brick_graphql_generators/src/graphql_serdes_generator.dart';
 import 'package:brick_json_generators/json_deserialize.dart';