+++ packages/brick_graphql_generators/test/graphql_model_serdes_generator/test_constructor_member_field_mismatch.dart
@@ -1,4 +1,4 @@
-import 'package:brick_graphql/graphql.dart';
+import 'package:brick_graphql/brick_graphql.dart';
 
 final output = r'''
 Future<GraphqlConstructorMemberFieldMismatch>