+++ packages/brick_graphql_generators/test/graphql_model_serdes_generator/test_unserializable_field_with_generator.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 import 'dart:io';
 
-import 'package:brick_graphql/graphql.dart';
+import 'package:brick_graphql/brick_graphql.dart';
 
 final output = r'''
 Future<GraphqlUnserializableFieldWithGenerator>