+++ packages/brick_offline_first_with_graphql_build/test/offline_first_generator/test_graphql_config_field_rename.dart
@@ -1,4 +1,4 @@
-import 'package:brick_graphql/graphql.dart' show FieldRename, GraphqlSerializable;
+import 'package:brick_graphql/brick_graphql.dart' show FieldRename, GraphqlSerializable;
 import 'package:brick_offline_first_with_graphql/brick_offline_first_with_graphql.dart';
 import 'package:brick_offline_first/brick_offline_first.dart';